# backend/openapi/agroconnect.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AgroConnect.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/farmer/me": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetMyFarm",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyFarmResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/prices": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetPrices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/weather": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetWeather",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/crop-check": {
      "post": {
        "tags": [
          "Farmer"
        ],
        "operationId": "CheckCrop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropCheckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/harvest-forecast": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetHarvestForecast",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarvestForecastResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/cooperative": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetCooperative",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooperativeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/lessons": {
      "get": {
        "tags": [
          "Farmer"
        ],
        "operationId": "GetLessons",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/farmer/change-requests": {
      "post": {
        "tags": [
          "Farmer"
        ],
        "operationId": "RequestChange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeArea": {
        "enum": [
          "phone",
          "farm",
          "crops",
          "other"
        ]
      },
      "ChangeRequest": {
        "required": [
          "area",
          "details"
        ],
        "type": "object",
        "properties": {
          "area": {
            "$ref": "#/components/schemas/ChangeArea"
          },
          "details": {
            "type": "string"
          }
        }
      },
      "ChangeRequestResponse": {
        "required": [
          "source",
          "reference"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "reference": {
            "type": "string"
          }
        }
      },
...
      "CooperativeResponse": {
        "required": [
          "source",
          "name",
          "community",
          "members",
          "chairName",
          "chairPhoneE164",
          "nextMeeting",
          "meetingPlace"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "name": {
            "type": "string"
          },
          "community": {
            "type": "string"
          },
          "members": {
            "type": "integer",
            "format": "int32"
          },
          "chairName": {
            "type": "string"
          },
          "chairPhoneE164": {
            "type": [
              "null",
              "string"
            ]
          },
          "nextMeeting": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "meetingPlace": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CropCheckRequest": {
        "required": [
          "crop",
          "symptoms"
        ],
        "type": "object",
        "properties": {
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "symptoms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CropSymptom"
            }
          }
        }
      },
      "CropCheckResponse": {
        "required": [
          "source",
          "likelyProblem",
          "urgent",
          "advice"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "likelyProblem": {
            "$ref": "#/components/schemas/CropProblem"
          },
          "urgent": {
            "type": "boolean"
          },
          "advice": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CropAdvice"
            }
          }
        }
      },
...
      "CropPrice": {
        "required": [
          "crop",
          "markets",
          "weekChangePercent",
          "last30Days"
        ],
        "type": "object",
        "properties": {
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "markets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketPrice"
            }
          },
          "weekChangePercent": {
            "type": "number",
            "format": "double"
          },
          "last30Days": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
...
      "DataSource": {
        "enum": [
          "sample",
          "live"
        ]
      },
...
      "FarmerProfile": {
        "required": [
          "id",
          "fullName",
          "phoneE164",
          "hasNoPhone",
          "gender",
          "ageBand",
          "community",
          "regionDistrict",
          "language",
          "consentAt",
          "crops",
          "farmSize",
          "farmSizeUnit",
          "soil",
          "plantingSeasons",
          "latitude",
          "longitude",
          "locationAccuracyMetres",
          "phoneType",
          "reachChannels",
          "mobileMoney",
          "helpNeeded",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "phoneE164": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasNoPhone": {
            "type": "boolean"
          },
          "gender": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "ageBand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgeBand"
              }
            ]
          },
          "community": {
            "type": [
              "null",
              "string"
            ]
          },
          "regionDistrict": {
            "type": [
              "null",
              "string"
            ]
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "consentAt": {
            "type": "string",
            "format": "date-time"
          },
          "crops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Crop"
            }
          },
          "farmSize": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "farmSizeUnit": {
            "$ref": "#/components/schemas/AreaUnit"
          },
          "soil": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SoilType"
              }
            ]
          },
          "plantingSeasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantingSeason"
            }
          },
          "latitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "longitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "locationAccuracyMetres": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "phoneType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PhoneType"
              }
            ]
          },
          "reachChannels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactChannel"
            }
          },
          "mobileMoney": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MobileMoneyUse"
              }
            ]
          },
          "helpNeeded": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HelpNeed"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HarvestForecastResponse": {
        "required": [
          "source",
          "crops"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "crops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CropForecast"
            }
          }
        }
      },
...
      "LessonsResponse": {
        "required": [
          "source",
          "lessons"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lesson"
            }
          }
        }
      },
...
      "MyFarmResponse": {
        "required": [
          "farmer",
          "officer",
          "visits"
        ],
        "type": "object",
        "properties": {
          "farmer": {
            "$ref": "#/components/schemas/FarmerProfile"
          },
          "officer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OfficerContact"
              }
            ]
          },
          "visits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VisitSummary"
            }
          }
        }
      },
      "OfficerContact": {
        "required": [
          "fullName",
          "phoneE164",
          "district"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phoneE164": {
            "type": "string"
          },
          "district": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PricesResponse": {
        "required": [
          "source",
          "updatedAt",
          "currency",
          "markets",
          "prices"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": "string"
          },
          "markets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CropPrice"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "VisitSummary": {
        "required": [
          "id",
          "status",
          "scheduledFor",
          "completedAt",
          "topics",
          "observations",
          "notes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/VisitStatus"
          },
          "scheduledFor": {
            "type": "string",
            "format": "date"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VisitTopic"
            }
          },
          "observations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmObservation"
            }
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "WeatherAdvice": {
        "enum": [
          "good_to_spray",
          "rain_soon_dont_spray",
          "dry_spell_water",
          "storm_protect_harvest"
        ]
      },
...
      "WeatherResponse": {
        "required": [
          "source",
          "updatedAt",
          "place",
          "today",
          "nextDays",
          "advice"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "place": {
            "type": "string"
          },
          "today": {
            "$ref": "#/components/schemas/WeatherDay"
          },
          "nextDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeatherDay"
            }
          },
          "advice": {
            "$ref": "#/components/schemas/WeatherAdvice"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The token from POST /api/auth/verify.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}